Change the logic for when we sometimes don't run incremental namer
The comment in the implementation below describes more fully.

You can read more about the motivation for this change in the commit for the PR that added it:

<https://github.com/sorbet/sorbet/pull/5989/commits/78b8f6f9b2b659b0da7a0c6f15cc7f20412e0142>

Simply deleting the check added in that commit still didn't work,
because the `CanPreemptSlowPathWithCodeAction` test still failed.
(It seems the changes we made in #6176 to skip re-running infer if it
has already been typechecked on a newer epoch did not apply, because the
epoch for a retypecheck command is always manually set to `0`, so it
looked like the last type check was older, not newer than the slow path
that might have been pre-empted. I think that epoch of `0` mostly makes
sense because we wouldn't want any of the errors that a `retypecheck`
produces to ever get reported. Only the errors in response to real edits
should get reported. And we use epochs to decide which is the latest
version of a file to report errors for.)

But as for the problem alluded to in the original commit, the other half
of the change was fixed; fast path edits can once again pre-empt slow
path edits (that is #6176), which means that the `.empty()` check was
actually over eager, and we can be okay handling all fast path edits
using incremental namer now.